backend/requirements.txt
fastapi
uvicorn
python-multipart==0.0.9
python-dotenv==1.0.1
pdfplumber==0.11.4
thefuzz
groq==0.11.0
supabase==2.7.2
apscheduler==3.10.4
pandas==2.2.2
httpx==0.27.2
sentence-transformers
torch
